remove int cast from register method
git-svn-id: http://scm.dspace.org/svn/repo/trunk@1553 9c30dcfa-912a-0410-8fc2-9e0234be79fd

+++ dspace/src/org/dspace/storage/bitstore/BitstreamStorageManager.java
@@ -499,7 +499,7 @@ public class BitstreamStorageManager
 		}
 
 		bitstream.setColumn("checksum_algorithm", "MD5");
-		bitstream.setColumn("size_bytes", (int) file.length());
+		bitstream.setColumn("size_bytes", file.length());
 		bitstream.setColumn("deleted", false);
 		DatabaseManager.update(context, bitstream);
